xds: throw away subchannel references after ring_hash is shutdown (#8140)
Similar to 368c43aec4bbf73999fab8ab05edddbd7a4a2b0d. 

Clean up subchannels after the RingHashLoadBalancer itself is shutdown to prevent further balancing state updates being propagated to the upstream.

Note this should not be considered as a fix for any problem anybody is noticing. Upstreams of RingHashLoadBalancer should not rely on this, it should still have its own logic for maintaining the lifecycle of downstream LB and ignore invalid upcalls when necessary.